mv install logic to internal/install
